chore(EP03): complete
From this episode, I had to come up with a workaround with the camera
not properly tracking the window on resize. I think this is likely just a
mistake with the camera implementation? Or maybe this is expected
behavior? I am not sure.

Some more rehashing of adding systems, queries, and also basic player
movement with the Input Resource.

Questions:
- What is an "entity" and how is is different than a bundle?
- How are "resources" made? Is there a listing of all of them?